Move /user/activate/<payload> route into users
Move supporting functions as well:
  - get_serializer
  - get_activation_link
